Modularize svg export
